Fix issue where opening a new overlay while closing a first.
When closing one row info dialog overly by opening a second, w2ui entered into an endless loop of checking the location of the popup and animating a transition between the two locations.  Fixed the way the checkIfMoved function is being called, and removed a callback to checkIfMoved that was causing the check to continue even if all dialogs were closed.